wth - (Fulfillment) Participant Tracker: Stop Defaulting Service
Background: In Fulfillment Participant Tracker, after chosen a visit, at
the bottom of the page, there is a dropdown box for the clinical
providers to be able to add more services (procedures). Currently, it is
always defaulting to be the first service of the dropdown box.

1). Changed the default to "Nothing Selected", so that the users won't
accidentally add the first service;
and
2). Added an error message for "Please choose the service that is desired
to be added." when there is no service chosen when "Add" is clicked.

[#145460053]

Story - (https://www.pivotaltracker.com/story/show/145460053)