wth - (SPARCFulfillment) Make the Klok Import Resource Format Rule Stricter
Please make the importing rule for Klok resource to be stricter, so that the ones do not follow the format of  ******(*** ) would error out and also cause the Klok entries to error out and show up in the proof report and import report.

Currently, klok/person.rb is extracting NetID from people.name inside the brackets, however, when there are no brackets, the last word of people.name is used as the NetID. That's causing issues because there could be a lot of people with the same last name and we are grabbing the wrong person for fulfillment.

[#147196653]

Story - https://www.pivotaltracker.com/story/show/147196653