wth - (SPARCFulfillment) Order the List of Clinical Providers
When assigning tasks to clinical providers from the "All Task" tab in SPARCFulfillment, or assign a clinical provider for the followup procedure, the list is out of order and showing duplicated people, as shown in attached screenshots.

Please 1). make the display of clinical providers unique;
2). display the list of names alphabetically.

[#152158482]

Story - https://www.pivotaltracker.com/story/show/152158482