wth - (SPARCFulfillment) Users should be Able to Delete Individual Fulfillment
Scenario: When a user has a non-clinical service wrongly fulfilled in SPARCFulfillment, there is currently no way to delete the individual fulfillment, or the service (because the validation checks whether the line item has fulfillments on it).

Please add the functionality on the fulfillment popup window "Activity" dropdown menu, for users to be able to delete an individual fulfillment, (and do NOT change the logic for validation).

[#154247222]

Story - https://www.pivotaltracker.com/story/show/154247222