Check admin_rates, then admin_rate_changes
First check admin_rates table for an existing record and, if it was
created on or before the day of fufillment, use it. If not, check
admin_rate_changes for the latest record on or before date of
fulfillment and, if its cost_reset flag is false, use it. For everything
else, use Service#cost.