Update modified rate column calculation
In v3.5.0 we added full support for checking for a modified rate when stamping service_cost
on fulfillments. For existing fulfillments that had a modified rate that
wasn't accounted for when their cost was set, we wan't their modified
rate column to say "no". This patch achieves that by comparing the cost
to the regular pricing map costs before labeling it a modified rate.